safety-detection-system/backend/requirements.txt
后端Python依赖列表
使用 'pip install -r requirements.txt' 命令进行安装
fastapi
uvicorn[standard]
python-multipart
ultralytics
numpy
opencv-python
Pillow
websockets